semanas

+++ src/components/Entregas/EntregasComponent.js
@@ -2,47 +2,67 @@ import React, {useState,  useEffect}from 'react';
 import firebase from "../../instances/firebase";
 import { v4 as uuid } from 'uuid';
 import { Jumbotron } from 'reactstrap';
+import { connect } from 'react-redux';
 import './EntregasComponent.css';
 
-export default function Entrega(props) {
+function Entrega(props) {
+    console.log({props})
+    const actividadId = props.match.params.id
+
     const [pdfUrl, setPdfUrl] = useState([]);
+    const [firebaseInited, setfirebaseInited] = useState(false)
+
     const readFile = async(e) =>{
         const file = e.target.files[0];
         const id = uuid();
+        const userId = props.userLoggedIn.localId
         const storageRef = firebase.storage().ref("pdf").child(id);
-        const pdfRef=firebase.database().ref("Tareas").child("Hoy").child(id);
+        const pdfRef=firebase.database().ref(`users/${userId}/tareas/${actividadId}/${id}`)
         await storageRef.put(file);
         storageRef.getDownloadURL().then((url)=> {
             pdfRef.set(url);
+            console.log({urlurl:url})
             const newState = [...pdfUrl, { id, url }];
             setPdfUrl(newState);
         });
     }
 
     const getPdfUrl = () => {
-        const pdfRef = firebase.database().ref("Tareas").child("Hoy");
-        pdfRef.on('value', (snapshot) => {
-          const pdfUrls = snapshot.val();
-          const urls = [];
-          for (let id in pdfUrls) {
-            urls.push({ id, url: pdfUrls[id] });
-          }
-          const newState = [...pdfUrl, ...urls];
-          setPdfUrl(newState);
-        });
+        if (!firebaseInited && props.userLoggedIn.localId){
+            setfirebaseInited(true);
+            const userId = props.userLoggedIn.localId;
+            console.log({userLogg:props.userLoggedIn});
+            console.log({useridid:props.userLoggedIn.localId});
+
+            const pdfRef = firebase.database().ref(`users/${userId}/tareas/${actividadId}`)//.child().child('tareas').child(actividadId);
+                pdfRef.on('value', (snapshot) => {
+                const pdfUrls = snapshot.val();
+                console.log({pdfUrls})
+                const urls = [];
+                for (let id in pdfUrls) {
+                    urls.push({ id, url: pdfUrls[id] });
+                }
+                const newState = [...pdfUrl, ...urls];
+                setPdfUrl(newState);
+            });
+        }
+        
       };
     
     const deletePdf = (id) => {
-    const storageRef = firebase.storage().ref("pdf").child(id);
-    const pdfRef = firebase.database().ref("Tareas").child("Hoy").child(id);
-    storageRef.delete().then(() => {
-        pdfRef.remove();
-    });
+        if(firebaseInited){
+            const userId = props.userLoggedIn.localId;
+            const storageRef = firebase.storage().ref("pdf").child(id);
+            const pdfRef = firebase.database().ref(`users/${userId}/tareas/${actividadId}/${id}`)//.ref("Tareas").child("Hoy").child(id);
+            storageRef.delete().then(() => {
+                pdfRef.remove();
+            });
+        }
     };
 
     useEffect(() => {
         getPdfUrl();
-      }, []);
+      });
 
     return (
         <div>
@@ -80,3 +100,11 @@ export default function Entrega(props) {
         </div>
     );
 }
+
+const mapStateToProps = state => {
+    return {
+        userLoggedIn: state.authenticationStore.userLoggedIn
+    }
+  }
+
+  export default connect(mapStateToProps)(Entrega);
